Store public and private key on the field
We will want to use these two values to encrypt and decrypt the
serialised data stored in the database. We will stored the armoured
version and then use PostgreSQL to switch this into bytes.

The key pair used in the test are a random pair that I created soley
for this use.