Save serialised encrypted data to database
This code relies on the pgcrypto extension to PostgreSQL in order to do
the PGP encryption, so we need to make sure this has been installed. We
now have a migration that's job is to install this extension into the
database being used by django.

I've had to split apart the tests from the core jsonpgpfield namespace.
This is because as soon as I introduced migrations to the field (for
pgcrypto) it required the test models to also be setup through
migrations. The complexity of trying to condition test migrations from
real was enough that splitting the code seemed most sensible.

The change in testing seemed to have some effect on how the arguments to
the field class were getting deconstructed. It ended up with the public
key stored in the verbose_name kwarg when deconstructing and no sign of
the private. I switched the args to be kwargs into the class, following
the pattern I've seen elsewhere. This means that the keys are no
defaulted to None.

Part of the motivation of this library vs using something out of the
py2.7 compatible version of
https://github.com/coldmind/django-pgcrypto-fields, was that we could
through something nicer than an InternalError if the field wasn't
instanciated with both keys. The code now throws out a NoKeyError when
trying to encrypt/save data when there is no public_key defined on the
instance.

In Python 2 errors should subclass from StandardError but this no longer
exists in Python 3 so I have introduced a dependency on six, a compat
library, so we can tell when in 3 or 2 and pick appropriately.